Only set last_commit_epoch if state is Commit
last_commit_epoch was getting set to the previous epoch regardless of
the state; this was a bug, because the definition of last_commit_epoch
is the epoch of the last Commit.

This fixes a bug where a participant can't recover because the
coordinator has essentially forgotten the relevant epoch. After this
patch, that error is recoverable by correcting the value of
last_commit_epoch in the context.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>