Add decision ack after commit/abort
Previously, the coordinator immediately went to the next epoch after
a decision was reached. This created a potential race condition in the
order of messages received and processed by a participant; if the next
epoch's RequestForVote message was received prior to the Commit/Abort
decision of the current epoch, then the RequestForVote message would be
dropped (because it wasn't for the current epoch). The result of the
dropped message was an eventual timeout and abort of that epoch.

This commit adds a new message sent from the participants to the
coordinator after a decision has been applied: DecisionAck. The
coordinator waits for all DecisionAck messages from all participants,
similar to collecting votes. Once all DecisionAcks are received, then
the coordinator starts the next epoch. If the coordinator times out, it
also starts the next epoch, since it is possible a decision ack will
never be sent by the processes that failed to respond.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>