Unset alarm when handling alarm for WaitingForStart/Vote
Previously, we handled the alarm and did not unset it. With this change,
we unset the alarm.

The algorithm doesn't set an alarm for these states. In the case of
WaitingForVote, no alarm is expected. Thus, it seems correct to unset
the alarm to match the algorithm's expectations.

The same is true for WaitingForStart, with the exception that we do
expect an Alarm to be the first event processed, in order to give the
alogorithm the chance to generate the correct initial notification(s),
in this case RequestForStart.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>